Query finding support.
Adde a module to make it easy for callers to store/find queries. Also
added a test program for the added module, updated the build process to
include the added module and updated the TODO to remove this item from
the TODO list.

Also updated the CHANGELOG to reflect this new feature.